Bandit protocol, classes and configuration file were added

+++ src/gossipLearning/models/bandits/BanditModel.java
@@ -4,6 +4,6 @@ import gossipLearning.interfaces.Model;
 
 public interface BanditModel extends Model {
   public double predict(int armIdx);
-  public long numberOfPlayes(int armIdx);
-  public long numberOfAllPlayes();
+  public double numberOfPlayes(int armIdx);
+  public double numberOfAllPlayes();
 }
